Hack-ish fix to get profiles to persist
It's bad for a couple reasons:
a) it duplicates the profile row instead
of overwriting each time the user saves a new one. But it's fine for a
demo because the profile does persist.

b) the profile store logic is in the user model and not the persistence
class, because I didn't code persistence with updating fields in mind.
This is a big TODO!!